requirements.txt
numpy
tifffile
pillow
vispy
pyqt